Moved placeShips method
